Implement support for client registration
Add beginnings of REST API definitions and support for registering as a
telemetry client, using a local instance id for identification purposes.
This will only be done if the system hasn't already been registered.

Added telemetry_base_url config option to the client config.

Added a testing config as testdata/config/localClient.yaml, which is
intended to work with a telemetry service instance running locally
on port 9999.