Misc enhancements to support server implementation
Processors and extractors only need to be provided with the datastore
configuration; this allows then to be used with either a client or a
server configuration.

Rename ReportToBundles to ReportsToBundles() since it processes all
available reports, not just one report.

Add DeleteDataItem to the TelemetryExtractor interface definition to
match the previously added implementation.

Add unit testing support for testing DeleteDataItem() as part of the
extractor tests.