Add support for telemetry client (re-)authenticate request
Define the over the wire representations for the /authenticate request
and response data structures. The request sends the existing clientId
and the SHA256 hash of the client's InstanceID, and gets back a similar
response to the /register request.

Add an Authenticate() method to the TelemetryClient that leverages the
new request to retrieve an update authentication bundle.

Add options to cmd/generator to allow control of whether registration
or authentication is performed.

Fix cmd/generator and cmd/clientds debug override setup to work
correctly.

Switch some slog messages from Info to Debug level.