Add cmd/authenticator and misc command tweaks
Add an authenticator command that will register, and re-authenticate,
a telemetry client system, and will report the current auth token's
issuer and expiration date.

Extended TelemtryClient with methods to get the ExpirationDate() and
Issuer() for the currently active token. Added internal helper methods
to support this.

Quietened some log messages by switching them to debug level and
tweaked some others to be more useful.

Added output messages to cmd/generator describing the steps that it
has performed. Similarly updated cmd/clientds to say that nothing
was found if the client datastore is empty.

Note that this change pulls support for JWT processing into the client
side library.