Rework client registration and clientId management (#53)
* Rework client registration and clientId management

Previously the clientId was the Id assigned by the upstream telemetry
service when a client was registered. This has been renamed to the
registrationId, along with the accompanying request header and the
associated fields in the registration response and re-authentication
request.

Previously a client system would generated a client instance id value
and store it in a file; this has now been renamed to a registration
value. Where previously the client instance id was just a single
string value derived from the timestamp when the value was generated,
it now consists of three values, namely a clientId as a UUID string,
an optional systemUUID (if available), and a UTC timestamp for when
the registration was generated.

Where previously this registrationId from the client credentials was
used as the clientId field in telemetry bundle and report structures,
we now will use the clientId value from the client's registration.

Updated docs to reflect these changes.

* Correct logic flaw in client registration handling

Need to ensure that the registration is loaded before checking for
the authentication creds.

Also correct some lingering references to client instance id.

* Further registration management enhancements

Primarily ensuring that any existing registrations are loaded before
attempting to generate a new one.

Also add support for handling an HTTP StatusConflict (409) in the case
of a duplicate registration attempt by retrying registering the client
once after triggering a regeneration of the client registration.

* Updated docs to further reflect client registration changes

* Update doc/telemetryclientregistration.md

Co-authored-by: Gabriel Silva Bueno <gabriel.bueno@suse.com>

---------

Co-authored-by: Gabriel Silva Bueno <gabriel.bueno@suse.com>