Implement a ManagedFile type (#55)
The ManagedFile type implements the new FileManager interface, and
provides support for CRUD operations as well as backup and advisory
locking.

Initial updates to the config file handling to leverage ManagedFile
to access the config file.

Also added support for display test coverage details after running
the tests in the test-verbose make target.