Authenication failures may require re-registration
In some cases an authentication failure may require a client to
re-register rather than just re-authenticate.

Migrate credentials management to use the ManagedFile mechanisms.
Also rename TelemetryAuth to TelemetryCreds, and similarly for
associated field in the TelemetryClient.

Use the ClientRegistrationResponse structure as the credientials
type, rather than replicating as an almost equivalent type.

Update registration and credentials backing types with validate
tags and use the validator to ensure that the contents are valid
after loading the content from disk.

Re-work the TelemetryClient initialisation to load the registration
and credentials files when they already exist.

Add Accessible() method to ManagedFile that can be used to confirm
that a file not only exists but can be read. Added a test to verify
this new method.

Also update DataItem, Bundle and Report structure json tags to
remove omitempty directives that conflict with validate required
directives.