Authenication failures may require re-registration (#62)
* Authenication failures may require re-registration

In some cases an authentication failure may require a client to
re-register rather than just re-authenticate.

Migrate credentials management to use the ManagedFile mechanisms.
Also rename TelemetryAuth to TelemetryCreds, and similarly for
associated field in the TelemetryClient.

Use the ClientRegistrationResponse structure as the credientials
type, rather than replicating as an almost equivalent type.

Update registration and credentials backing types with validate
tags and use the validator to ensure that the contents are valid
after loading the content from disk.

Re-work the TelemetryClient initialisation to load the registration
and credentials files when they already exist.

Add Accessible() method to ManagedFile that can be used to confirm
that a file not only exists but can be read. Added a test to verify
this new method.

Also update DataItem, Bundle and Report structure json tags to
remove omitempty directives that conflict with validate required
directives.

* Skip Accessible checks if running as root

Split the Accessible test out into a separate test that only runs if
running as a non-root user.

* Run tests as non-root user

* Use variable to hold test user account name

* Make dataitem footer and checksum optional

* Fix Valdiate typo