Add telemetry client tools container (#67)
Add a telemetry-tools container that contains the telemetry client
tools and the example app compiled and ready to use.

The Dockerfile builds a container image that includes the existing
telemetry client tools (generator, clientds, authenticator) as well
as the example app.

The entrypoint script will:
  * add any certs that are found in the directory specified by the
    TELEMETRY_CERTS_DIR env var to the container's runtime certs
    store, running update-ca-certificates as needed.
  * generate the config file specified by the TELEMTRY_CONFIG env var
    if not found, using the settings specified in the TELEMETRY_* env
    vars.
  * run the specified tool, automatically adding an appropriate
    minimum set of arguments for that tool along with any arguments
    that may have been specified for the tool by the user.

Add a testing telemetry blob.json that will be used as the default
telemetry blob when running the generator tool via the container.

Add support for specifying :memory: as the local datastore, rather
than a backing file. Add support for checking of the configured
datastore is persistent or not, ensuring that telemetry is submitted
rather that staged locally, when a non-persistent datastore is used.

Complete the implementation of telemetry submission in the library
code.

Enhance the example app to leverage the complete set of the library
features.

Update the data type of the Tags field in the client config to use
our Tags type.

Add Makefile support for building the new telemetry-tools image with
the docker-build target, as well as using this new image to generate
telemetry submissions to a server using the docker-generate target.