Update client registration hash generation
Given that JSON encoding is not always consistent across different
platforms, switch to using a custom encoding for ClientRegistration
for the input value when generating the ClientRegistrationHash.

Provide a HashJSON() method that uses the JSON encoding as input to
allow for backwards compatibility implementations.

Update documentation for the /authenticate request to document how
the regHash field is encoded, and the switch to using a platform
independent encoding scheme for the hashing input value.